+++ src/NServiceBus.Core/Transports/Msmq/MsmqReceiveWithNoTransactionBehavior.cs
@@ -33,59 +33,4 @@ namespace NServiceBus
             onMessage(transportMessage);
         }
     }
-
-
-    class MsmqReceiveWithNativeTransactionBehavior : MsmqReceiveBehavior
-    {
-        protected override void Invoke(IncomingContext context, Action<TransportMessage> onMessage)
-        {
-            var queue = context.Get<MessageQueue>();
-
-            using (var msmqTransaction = new MessageQueueTransaction())
-            {
-                try
-                {
-                    msmqTransaction.Begin();
-
-                    Message message;
-
-                    if (!TryReceiveMessage(() => queue.Receive(TimeSpan.FromSeconds(1), msmqTransaction), context, out message))
-                    {
-                        msmqTransaction.Commit();
-                        return;
-                    }
-
-                    TransportMessage transportMessage;
-
-                    try
-                    {
-                        transportMessage = MsmqUtilities.Convert(message);
-                    }
-                    catch (Exception ex)
-                    {
-                        HandleCorruptMessage(context, message, ex, (q, m) => q.Send(m, msmqTransaction));
-
-                        msmqTransaction.Commit();
-                        return;
-                    }
-
-                    //todo: add tx to UoW
-                    onMessage(transportMessage);
-                    //finally
-                    //{
-                    //    //todo: clear tx to UoW
-                    //}
-                    msmqTransaction.Commit();
-                }
-                catch (Exception)
-                {
-                    msmqTransaction.Abort();
-                }
-           }
-
-
-
-
-        }
-    }
 }
